fix: 개인정보 수정 null 대응, 리팩토링
